增加Controlled-UDP模块

+++ P2PerLib/src/ml/liuchibing/P2PerLib/SharedCodes.java
@@ -26,6 +26,14 @@ public class SharedCodes {
 
     public static final String REGEX_PACKET_HAS_CONTENT = "\\{.+\\}";
 
+    /**
+     * 发送Control数据.
+     * @param command
+     * @param content
+     * @param destination
+     * @param socket
+     * @throws IOException
+     */
     public static void SendControlData(String command, String content, InetSocketAddress destination, DatagramSocket socket) throws IOException {
         //构建Control字符串
         StringBuilder sb = new StringBuilder(SharedCodes.CURRENT_VERSION_HEAD);
@@ -39,7 +47,7 @@ public class SharedCodes {
         //构建buffer
         ByteArrayOutputStream buffer = new ByteArrayOutputStream();
         DataOutputStream dos = new DataOutputStream(buffer);
-        dos.writeInt(PacketType.Control.ordinal());//写入包类型
+        dos.writeByte((byte)PacketType.Control.ordinal());//写入包类型
         byte[] data = sb.toString().getBytes("utf-8");//编码数据
         dos.write(data, 0, data.length);//写入数据
         socket.send(new DatagramPacket(buffer.toByteArray(), buffer.toByteArray().length, destination));
